label letter mode add

+++ native/cocos/2d/renderer/Batcher2d.h
@@ -115,7 +115,7 @@ private:
         float_t* vbBuffer = entity->getVbBuffer();
 
         Vec3 temp;
-        uint8_t offset = 0;
+        uint32_t offset = 0;
         for (int i = 0; i < size; i += stride) {
             //Render2dLayout* curLayout = dataList[i];
             Render2dLayout* curLayout = entity->getRender2dLayout(i);
